tests: run kernel config check earlier
The runtime Xenomai tests are likely to lock and reboot the board before
the kernel config check is run. This is a pain because often one has to
comment these tests manually to get the kernel config results.

Print the results early so they always run, regardless of the Xenomai
stress tests.

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>